+++ node_app/views/flip.ejs
@@ -20,10 +20,13 @@
         </div>
         <div>
             <br><br><br>
-            <% if(articles != 0) { %>
+            <% if(response) { %>
                 <div>
                     <h1>Here is your flipped text:</h1>
                     <p><%= response %></p>
+                    <h2> Here are the words that were changed: </h2>
+                    <p><%= flippedArray  %> </p>
+                    
                     <%= articles = [] %>
                 </div>
             <% } %>
